Fix for #134. Fix version on tragl.

+++ _scripts/set-version.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/bash
-version="0.17.1"
+version="0.18.0"
 cd src
 directories=`find . -maxdepth 1 -type d -name "tr*"`
 cwd=`pwd`
@@ -40,7 +40,44 @@ do
 	mv asdfasdf *.csproj	
 	rm -f asdfasdf2
 	touch readme.md
-	cat readme.md | sed -e 's%^0[.][0-9]*[.][0-9]*[ ]*[-][-].*$'"%$version -- Fixes for all tools (piped data structures), but in particular trparse. Add trperf, trpiggy.%" > asdfasdf2
+	cat readme.md | sed -e 's%^0[.][0-9]*[.][0-9]*.*$'"%$version"' Fix 134 for all tools. Fix 180, string-length() and math operations in XPath engine. Fix for crash in https://github.com/antlr/grammars-v4/issues/2818 where _grammar-test was removed, but pom.xml not completely cleaned up of the reference to the directory. Fix Globbing package because of conflict with Antlr4BuildTasks. Update Antlr4BuildTasks version. Rename TreeEdits.Delete() so there is no confuson that it modifies entire parse tree, token and char streams. Add -R option for rename map as file in trrename. Update trrename to use xpath and TreeEdits. Add methods to TreeEdits. Rewrote trrename to use xpath/treeedits packages.%' > asdfasdf2
+	mv asdfasdf2 readme.md
+	cd ..
+done
+
+for i in tragl
+do
+	echo $i
+	if [ "$i" == "." ]
+	then
+		continue
+	fi
+	cd $cwd/$i
+	if [[ "$?" != "0" ]]
+	then
+		continue
+	fi
+	csproj=`find . -maxdepth 1 -name '*.csproj'`
+	if [[ "$csproj" == "" ]]
+	then
+		cd ..
+		continue
+	fi
+	if [[ ! -f "$i.csproj" ]]
+	then
+		continue
+	fi
+	if [[ ! -f "readme.md" ]]
+	then
+		continue;
+	fi
+	echo $i
+	rm -f asdfasdf
+	cat *.csproj | sed -e "s%[<][Vv]ersion[>].*[<][/][Vv]ersion[>]%<Version\>$version</Version>%" > asdfasdf
+	mv asdfasdf *.csproj	
+	rm -f asdfasdf2
+	touch readme.md
+	cat readme.md | sed -e 's%^0[.][0-9]*[.][0-9]*.*$'"%$version"' Fix 134 for all tools. Fix 180, string-length() and math operations in XPath engine. Fix for crash in https://github.com/antlr/grammars-v4/issues/2818 where _grammar-test was removed, but pom.xml not completely cleaned up of the reference to the directory. Fix Globbing package because of conflict with Antlr4BuildTasks. Update Antlr4BuildTasks version. Rename TreeEdits.Delete() so there is no confuson that it modifies entire parse tree, token and char streams. Add -R option for rename map as file in trrename. Update trrename to use xpath and TreeEdits. Add methods to TreeEdits. Rewrote trrename to use xpath/treeedits packages.%' > asdfasdf2
 	mv asdfasdf2 readme.md
 	cd ..
 done
